fixed an issue that was preventing an employee removal when the employee had no employee info associated. This was possible with employees with secondments from other PYSDE

+++ gr.sch.ira.minoas/src/main/java/gr/sch/ira/minoas/seam/components/home/EmployeeHome.java
@@ -291,7 +291,9 @@ public class EmployeeHome extends MinoasEntityHome<Employee> {
 				getEntityManager().remove(employee.getRegularDetail());
 			}
 			
-			getEntityManager().remove(employee.getEmployeeInfo());
+			if(employee.getEmployeeInfo()!=null) {
+				getEntityManager().remove(employee.getEmployeeInfo());
+			}
 			return super.remove();
 		}
 		return null;
